% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D3_gls.R
\name{D3_gls}
\alias{D3_gls}
\title{Compare two nested models using MI LRT}
\usage{
D3_gls(model, null_model, datalist)
}
\description{
The MI LRT (D3) test by Meng and Rubin (1992).
}
\references{
Meng, X. L., and D. B. Rubin. 1992.
Performing Likelihood Ratio Tests with Multiply-Imputed Data Sets.
\emph{Biometrika}, 79 (1): 103-11.
}
