% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchangeable_cor.R
\name{exchangeable_cor}
\alias{exchangeable_cor}
\title{Gives the compound symmetric correlation matrix}
\usage{
exchangeable_cor(n, rho)
}
\value{
A compound symmetric correlation matrix of order n.
}
\description{
(This is an internal function).
}
