% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pooled_estimates.R
\name{get_pooled_estimates}
\alias{get_pooled_estimates}
\title{Computes pooled estimates from the list containing `m` fitted models.}
\usage{
get_pooled_estimates(model)
}
\value{
A list containing the completed-data estimates and
their variance-covariance matrices.
}
\description{
(This is an internal function).
}
