% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rho.R
\name{get_rho}
\alias{get_rho}
\title{Gives the correlation value for compound symmetry correlation structure}
\usage{
get_rho(object)
}
\value{
A correlation value
}
\description{
(This is an internal function).
}
