% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_at_fixed_psi.R
\name{likelihood_at_fixed_psi}
\alias{likelihood_at_fixed_psi}
\title{Computes likelihood at fixed value of parameters.}
\usage{
likelihood_at_fixed_psi(object, psi)
}
\value{
The likelihood value at a given psi.
}
\description{
(This is an internal function).
}
